list of terms with bids for event

+++ src/main/java/org/example/beatmybet/entity/Bid.java
@@ -28,7 +28,7 @@ public class Bid implements Serializable, GlobalFinanceEntity {
     private Date date;
 
     @Column(name = "koef")
-    private Float koef;
+    private Double koef;
 
     @ManyToOne
     @JoinColumn(name = "id_var")
